Tested against linux and windows.

+++ scriptextension/start-csextension.ps1
@@ -2,7 +2,9 @@ function start-extension
     {
         [cmdletbinding()]
         param(
-            [parameter (Mandatory=$true)][psobject]$Target,
+            [parameter (Mandatory=$true)][string]$name,
+            [parameter (Mandatory=$true)][string]$ResourceGroupName,
+            [parameter (Mandatory=$false)][switch]$nofile,
             [parameter (Mandatory=$false)][string]$uri,
             [parameter (Mandatory=$true)][string]$ExtensionName,
             [parameter (Mandatory=$true)][string]$ExtensionType,
@@ -14,30 +16,55 @@ function start-extension
 
         try {
 
-            Set-AzureRmContext -Subscription $SubscriptionID -ErrorAction Stop
-            $vm = Get-AzureRmVm -ResourceGroupName $Target.ResourceGroupName -ResourceGroupName -Name $Target.Name
+            #Set-AzureRmContext -Subscription $SubscriptionID -ErrorAction Stop
+            $vm = Get-AzureRmVm -ResourceGroupName $ResourceGroupName -Name $name -ErrorAction Stop
+            Write-Host $VM
             $customScriptExtensions = $($vm.Extensions) | where {$_.VirtualMachineExtensionType -eq $ExtensionType}
             
-            try #change to loop that checks status Dont want to remove if in progress....
+            if($customScriptExtensions)
                 {
-                    $removeParms = @{
 
+                    $removeParms = @{
                         "VMName" = $($vm.Name);
                         "ResourceGroupName" =  $($vm.ResourceGroupName);
                         "Name" =  $($customScriptExtensions.Name);
                         "ErrorAction" = "stop"
+                        }
+                    
+                    try{    
+                        Remove-AzureRmVMExtension @removeParms -Force
                     }
                     
-                    Remove-AzureRmVMExtension @removeParms -Force
-
-                    $PublicConfiguration = @{
-                        "fileUris" = [Object[]]$uri;
-                        "timestamp" = $timestamp;
-                        "commandToExecute" = $command
+                    catch{
+                        Write-Host
+                        Write-Host "$($Error[0])" -BackgroundColor Red
+                        Write-Host
                     }
+
+
+                }
+
+            try #change to loop that checks status Dont want to remove if in progress....
+                {
+
+                    if($nofile = $false){
+                        $PublicConfiguration = @{
+                            "fileUris" = [Object[]]$uri;
+                            "timestamp" = $timestamp;
+                            "commandToExecute" = $command
+                        }
+
+                    }
+                    else{
+                        $PublicConfiguration = @{
+                            "timestamp" = $timestamp;
+                            "commandToExecute" = $command
+                        }
+                    }
+                    
                     
                     $CSEParams = @{
-                        "ResourceGroupName" = "$($vm.ResourceGroupName)";
+                        "ResourceGroupName" = "$($ResourceGroupName)";
                         "VMName" = "$($vm.Name)";
                         "Location" = "$($vm.Location)";
                         "Name" = $ExtensionName;
@@ -51,7 +78,7 @@ function start-extension
                     try {
                         Set-AzureRmVMExtension @CSEParams
                         #may need to put next section in a do-while.
-                        $VMstatus = Get-AzureRmVm -Name $vm.Name -ResourceGroupName $vm.ResourceGroupName -status | where {$_.Name -eq $ExtensionName}
+                        $VMstatus = Get-AzureRmVm -Name $vm.Name -ResourceGroupName $ResourceGroupName -status | where {$_.Name -eq $ExtensionName}
                         return $VMstatus.Extensions.Statuses.DisplayStatus
                     } 
                     
@@ -69,7 +96,6 @@ function start-extension
                 }       
         }
         catch {
-
             Write-Host
             Write-Host "$($Error[0])" -BackgroundColor Red
             Write-Host
